backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-modeltranslation==0.19.14
django-rest-passwordreset==1.5.0
django-sslserver==0.22
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
idna==3.10
openpyxl==3.1.5
payme-pkg==3.0.24
pillow==11.2.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==12.1.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
